# sprint_mission_4_base/postman/specs/Discodeit API 문서/api-docs-postman-openapi3.0.3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Discodeit API 문서",
    "description": "Discodeit 프로젝트의 Swagger API 문서입니다.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 서버"
    }
  ],
...
  "paths": {
...
    "/api/readStatuses": {
      "get": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "User의 Message 읽음 상태 목록 조회",
        "operationId": "findAllByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "조회할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 읽음 상태 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReadStatus"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "Message 읽음 상태 생성",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadStatusCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음",
            "content": {
              "application/json": {
                "example": "Channel | User with id {channelId | userId} not found"
              }
            }
          },
          "400": {
            "description": "이미 읽음 상태가 존재함",
            "content": {
              "application/json": {
                "example": "ReadStatus with userId {userId} and channelId {channelId} already exists"
              }
            }
          },
          "201": {
            "description": "Message 읽음 상태가 성공적으로 생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStatus"
                }
              }
            }
          }
        }
      }
    },
    "/api/messages": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Channel의 Message 목록 조회",
        "operationId": "findAllByChannelId",
        "parameters": [
          {
            "name": "channelId",
            "in": "query",
            "description": "조회할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Message 생성",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageCreateRequest": {
                    "$ref": "#/components/schemas/MessageCreateRequest"
                  },
                  "attachments": {
                    "type": "array",
                    "description": "Message 첨부 파일들",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "messageCreateRequest"
                ]
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음",
            "content": {
              "application/json": {
                "example": "Channel | Author with id {channelId | authorId} not found"
              }
            }
          },
          "201": {
            "description": "Message가 성공적으로 생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/readStatuses/{readStatusId}": {
      "patch": {
        "tags": [
          "ReadStatus"
        ],
        "summary": "Message 읽음 상태 수정",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "readStatusId",
            "in": "path",
            "description": "수정할 읽음 상태 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message 읽음 상태가 성공적으로 수정됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadStatus"
                }
              }
            }
          },
          "404": {
            "description": "Message 읽음 상태를 찾을 수 없음",
            "content": {
              "application/json": {
                "example": "ReadStatus with id {readStatusId} not found"
              }
            }
          }
        }
      }
    },
...
    "/api/binaryContents": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "여러 첨부 파일 조회",
        "operationId": "findAllByIdIn",
        "parameters": [
          {
            "name": "binaryContentIds",
            "in": "query",
            "description": "조회할 첨부 파일 ID 목록",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "첨부 파일 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BinaryContent"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReadStatusCreateRequest": {
        "type": "object",
        "description": "Message 읽음 상태 생성 정보",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReadStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "lastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MessageCreateRequest": {
        "type": "object",
        "description": "Message 생성 정보",
        "properties": {
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "ReadStatusUpdateRequest": {
        "type": "object",
        "description": "수정할 읽음 상태 정보",
        "properties": {
          "newLastReadAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "BinaryContent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          },
          "bytes": {
            "type": "string",
            "format": "byte"
          }
        }
      }
    }
  }
}